+++ CHANGELOG.md
@@ -10,6 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 - Homeassistant device info to map entities per device in the integration.
 - Debugging logs for API calls to be able to trace received data.
+- Missing Binary sensors (as reported by [killer8](https://tweakers.net/gallery/304893/)):
+  - Active WW request
+  - Direct Outlet (DG1)
+  - Mixture Outlet (SG2)
+  - Mixture Outlet (SG3)
+  - Mixture Outlet (SG4)
+
+
 
 ### Removed
 - Device name from default entity name